refactor: adapater response domain
