+++ front/src/modules/adapter/README.md
@@ -1,0 +1,1 @@
+レスポンスをドメインへ変換する
