+++ front/src/modules/adapter/top/domain.ts
@@ -1,8 +1,6 @@
-import { PageProps } from 'gatsby';
 import { DomainTopCategory, DomainTopCategoryBlog } from '@/domain/top/blog';
 import { DomainTopStoryBlog } from '@/domain/top/storyBlog';
 import { errorWrapper } from '@/modules/common/error';
-import { Res } from '@/modules/interfaces/response/home';
 import {
   ResTocoBlog,
   ResCategory,
@@ -10,31 +8,9 @@ import {
 import { ResStoryBlog } from '@/modules/interfaces/response/home/storyBlogs';
 
 /**
- * レスポンス取得: カテゴリ一覧
+ * レスポンスをドメインへ変換: ブログ一覧
  */
-const getResCategory = (r: Res): ResCategory[] => {
-  try {
-    return r.strapi.categories.data || ([] as ResCategory[]);
-  } catch (e) {
-    throw errorWrapper(e, 'レスポンス取得エラー');
-  }
-};
-
-/**
- * レスポンス取得: ストーリー記事一覧
- */
-const getResStoryBlogs = (r: Res): ResStoryBlog[] => {
-  try {
-    return r.strapi.storyBlogs.data || ([] as ResStoryBlog[]);
-  } catch (e) {
-    throw errorWrapper(e, 'レスポンス取得エラー:ストーリー記事一覧');
-  }
-};
-
-/**
- * レスポンスドメイン変換: ブログ一覧
- */
-const getDomainTopCategoryBlog = (
+export const getDomainTopCategoryBlog = (
   resblogs: ResTocoBlog[]
 ): DomainTopCategoryBlog[] => {
   try {
@@ -56,9 +32,9 @@ const getDomainTopCategoryBlog = (
 };
 
 /**
- * レスポンスドメイン変換: カテゴリ一覧
+ * レスポンスをドメインへ変換: カテゴリ一覧
  */
-const getDomainTopCategory = (
+export const getDomainTopCategory = (
   resCategories: ResCategory[]
 ): DomainTopCategory[] => {
   try {
@@ -80,9 +56,11 @@ const getDomainTopCategory = (
 };
 
 /**
- * レスポンスドメイン変換: ストーリー記事一覧
+ * レスポンスをドメインへ変換: ストーリー記事一覧
  */
-const getDomainStoryBlog = (blogs: ResStoryBlog[]): DomainTopStoryBlog[] => {
+export const getDomainStoryBlog = (
+  blogs: ResStoryBlog[]
+): DomainTopStoryBlog[] => {
   try {
     return blogs.map(r => {
       const id = Number(r.id) || 0;
@@ -97,38 +75,5 @@ const getDomainStoryBlog = (blogs: ResStoryBlog[]): DomainTopStoryBlog[] => {
     });
   } catch (e) {
     throw errorWrapper(e, 'ドメイン変換エラー:ストーリー記事一覧');
-  }
-};
-
-interface useReturn {
-  readonly categories: DomainTopCategory[];
-  readonly stroyBlogs: DomainTopStoryBlog[];
-}
-
-/**
- * トップページ
- * GraphQLのレスポンスをドメインに変換
- */
-export const adapterDomainTop = (page: PageProps): useReturn => {
-  try {
-    /**
-     * レスポンスを取得
-     */
-    const res = page.data as Res;
-    const resCategories: ResCategory[] = getResCategory(res);
-    const resStoryBlogs: ResStoryBlog[] = getResStoryBlogs(res);
-
-    /**
-     * ドメインに変換
-     */
-    const categories: DomainTopCategory[] = getDomainTopCategory(resCategories);
-    const stroyBlogs: DomainTopStoryBlog[] = getDomainStoryBlog(resStoryBlogs);
-
-    return {
-      categories,
-      stroyBlogs,
-    };
-  } catch (e) {
-    throw errorWrapper(e, 'レスポンス変換エラー');
   }
 };
